+++ ir/be/amd64/amd64_emitter.c
@@ -698,6 +698,7 @@ static void emit_amd64_asm_register(const arch_register_t *reg, char modifier,
 
 static void emit_amd64_asm_operand(ir_node const *const node, char const modifier, unsigned const pos)
 {
+	x86_asm_operand_kind_t required;
 	switch (modifier) {
 	case '\0':
 	case 'b':
@@ -705,6 +706,11 @@ static void emit_amd64_asm_operand(ir_node const *const node, char const modifie
 	case 'k':
 	case 'q':
 	case 'w':
+		required = ASM_OP_INVALID;
+		break;
+
+	case 'c':
+		required = ASM_OP_IMMEDIATE;
 		break;
 
 	default:
@@ -714,6 +720,13 @@ static void emit_amd64_asm_operand(ir_node const *const node, char const modifie
 
 	be_asm_attr_t     const *const attr = get_be_asm_attr_const(node);
 	x86_asm_operand_t const *const op   = &((x86_asm_operand_t const*)attr->operands)[pos];
+
+	if (required != ASM_OP_INVALID && required != op->kind) {
+		char const *const name = x86_get_constraint_name(required);
+		be_errorf(node, "asm modifier '%c' requires an operand of type '%s'", modifier, name);
+		return;
+	}
+
 	switch ((x86_asm_operand_kind_t)op->kind) {
 	case ASM_OP_INVALID:
 		panic("invalid asm operand");
@@ -737,7 +750,7 @@ static void emit_amd64_asm_operand(ir_node const *const node, char const modifie
 	}
 
 	case ASM_OP_IMMEDIATE: {
-		amd64_emit_immediate32(true, &op->u.imm32);
+		amd64_emit_immediate32(modifier != 'c', &op->u.imm32);
 		return;
 	}
 	}